Switch to a voluntary kernel preemption. The respeaker mic hats and possibly (many) others are using the x-powers ac108 adc chip however the driver of this chip is not compatible with 64bit PREEMPT kernels crashing the drivers with "BUG: scheduling while atomic" spinlock issues.
PREEMPT is used for low latency desktops
PREEMPT_NONE is used for servers

As we are not a (low-latency needed) desktop however not a server
as well, the middle ground actually works perfectly fine with
the benefit of having a more stable respeaker driver.

This accounts for ALL different respeaker HATs except the 2-mic
as that one is based on the WM8960 chip which is actually already
supported mainline if you want anyway.